Moving HttpClient related code into base package.
This allows us to make all these types package-private, and to not
expose them as public surface area.
